Reformatting and cleaning up
Reformatting Object file formations
Cleaning up unused codes
Renaming files to standardise
