+++ api/classs.py
@@ -1,0 +1,37 @@
+from main import db
+
+
+class Class(db.Model):
+    __tablename__ = 'classes'
+
+    clsCourseID = db.Column(db.String(30), db.ForeignKey('courses.courseID', ondelete="CASCADE", onupdate="CASCADE"), primary_key=True)
+    classID = db.Column(db.Integer, primary_key=True)
+    clsTrainer = db.Column(db.String(30), nullable=False)
+    clsStartTime = db.Column(db.String(30), nullable = False)
+    clsEndTime = db.Column(db.String(30), nullable=False)
+    clsLimit = db.Column(db.Integer, nullable=False)
+    regPeriod = db.Column(db.String(30), nullable=False)
+    # user = relationship('User', backref='child')
+    courseClass = db.relationship(
+    'Course', primaryjoin='Class.clsCourseID == Course.courseID', backref='classes')
+
+    def __init__(self, clsCourseID, classID, clsTrainer, clsStartTime, clsEndTime, clsLimit, regPeriod):
+        self.clsCourseID = clsCourseID
+        self.classID = classID
+        self.clsTrainer = clsTrainer
+        self.clsStartTime = clsStartTime
+        self.clsEndTime = clsEndTime
+        self.clsLimit = clsLimit
+        self.regPeriod = regPeriod
+        
+
+    def json(self):
+        return {
+        "clsCourseID": self.clsCourseID, 
+        "classID": self.classID, 
+        "clsTrainer": self.clsTrainer, 
+        "clsStartTime": self.clsStartTime, 
+        "clsEndTime": self.clsEndTime,
+        "clsLimit": self.clsLimit,
+        "regPeriod": self.regPeriod
+        }
