+++ api/completed.py
@@ -1,0 +1,23 @@
+from main import db
+
+
+
+class Completed(db.Model):
+    __tablename__ = 'completedCourses'
+
+    ccStudentID = db.Column(db.Integer, db.ForeignKey('students.studentID', ondelete="CASCADE", onupdate="CASCADE"), primary_key=True)
+    completedCName = db.Column(db.String(30), primary_key=True)
+
+    # user = relationship('User', backref='child')
+    studentCompleted = db.relationship(
+    'Student', primaryjoin='Student.studentID == Completed.ccStudentID', backref='completedCourses')
+
+    def __init__(self, ccStudentID, completedCName):
+        self.ccStudentID = ccStudentID
+        self.completedCName = completedCName
+    
+    def json(self):
+        return {
+        "ccStudentID": self.ccStudentID, 
+        "completedCName": self.completedCName
+        }
