+++ api/student.py
@@ -1,0 +1,23 @@
+from main import db
+
+
+class Student(db.Model):
+    __tablename__ = 'students'
+
+    studentID = db.Column(db.Integer, primary_key=True)
+    studentName = db.Column(db.String(30), nullable=False)
+    sPosition = db.Column(db.String(30), nullable = False)
+
+
+    def __init__(self, studentID, studentName, sPosition):
+        self.studentID = studentID
+        self.studentName = studentName
+        self.sPosition = sPosition
+
+
+    def json(self):
+        return {
+            "studentID" : self.studentID,
+            "studentName": self.studentName,
+            "sPosition" : self.sPosition,
+        }
