+++ api/testApp.py
@@ -128,52 +128,6 @@ class TestRegistration(TestApp):
         self.assertEqual(response.json['data'], {"regStudentID": 1, "regCourseID": "1", "regClassID": 3, "regStatus": "enrolled"}
         )
     
-    # def test_create_registration_already_taken(self):
-        # d1 = Course(1, '3D Printing Software v1.0', 'A course on 3D printing software', '3D Printing Basics, 3D Printer Software Installation', False)
-        # d2 = Class(1,3, "Lim Ah Hock", "12-Sept-2021", "14-Sept-2023", 35, "9 Oct, 2021 to 9 Nov, 2021")
-        # d3 = Registration(1,1,3,"enrolled")
-        # db.session.add(d1)
-        # db.session.add(d2)
-        # db.session.add(d3)
-        # db.session.commit()
-
-        # request_body = {
-        #     "regStudentID": 1, 
-        #     "regCourseID": 1, 
-        #     "regClassID": 3, 
-        #     "regStatus": "enrolled"
-        # }
-        # response = self.client.post("/registerClass",
-        #                             data=json.dumps(request_body),
-        #                             content_type='application/json')
-        # print(response.json)
-        # self.assertEqual(response.json['message'], {"Student has already been registered"}
-        # )
-
-    # def test_create_registration_prereqs_not_satisfied(self):
-        #need to test initialising data in setUp function instead so no need to copy-paste
-    
-        # d1 = Course(1, '3D Printing Software v1.0', 'A course on 3D printing software', '3D Printing Basics, 3D Printer Software Installation', False)
-        # d2 = Class(1,3, "Lim Ah Hock", "12-Sept-2021", "14-Sept-2023", 35, "9 Oct, 2021 to 9 Nov, 2021")
-        # d3 = Registration(1,1,3,"enrolled")
-        # db.session.add(d1)
-        # db.session.add(d2)
-        # db.session.add(d3)
-        # db.session.commit()
-
-        # request_body = {
-        #     "regStudentID": 1, 
-        #     "regCourseID": 1, 
-        #     "regClassID": 3, 
-        #     "regStatus": "enrolled"
-        # }
-        # response = self.client.post("/registerClass",
-        #                             data=json.dumps(request_body),
-        #                             content_type='application/json')
-        # print(response.json)
-        # self.assertEqual(response.json['message'], {"Student has already been registered"}
-        # )
-    # def test_create_registration_already_registered(self):
 
 class TestCompleted(TestApp):
